BACKPORT: FROMLIST: dma-buf/heaps: Correct the types of fd_flags and heap_flags
dma_heap_allocation_data defines the UAPI as follows:

 struct dma_heap_allocation_data {
        __u64 len;
        __u32 fd;
        __u32 fd_flags;
        __u64 heap_flags;
 };

But dma heaps are casting both fd_flags and heap_flags into
unsigned long. This patch makes dma heaps - cma heap and
system heap have consistent types with UAPI.

Acked-by: John Stultz <jstultz@google.com>

Bug: 344501512
Link: https://lore.kernel.org/lkml/20240606020213.49854-1-21cnbao@gmail.com/
[cmllamas: fixed trivial conflicts]

+++ include/linux/dma-heap.h
@@ -24,8 +24,8 @@ struct dma_heap;
 struct dma_heap_ops {
 	struct dma_buf *(*allocate)(struct dma_heap *heap,
 				    unsigned long len,
-				    unsigned long fd_flags,
-				    unsigned long heap_flags);
+				    u32 fd_flags,
+				    u64 heap_flags);
 	long (*get_pool_size)(struct dma_heap *heap);
 };
